rework based on feedback, move matcher into search package
move alias to separate package
