+++ cmd/ld-find-code-refs/main.go
@@ -8,7 +8,6 @@ import (
 	"github.com/launchdarkly/ld-find-code-refs/coderefs"
 	"github.com/launchdarkly/ld-find-code-refs/internal/log"
 	"github.com/launchdarkly/ld-find-code-refs/internal/version"
-	"github.com/launchdarkly/ld-find-code-refs/matcher"
 	o "github.com/launchdarkly/ld-find-code-refs/options"
 )
 
@@ -56,7 +55,7 @@ var cmd = &cobra.Command{
 		}
 
 		log.Init(opts.Debug)
-		matcher.Run(opts)
+		coderefs.Run(opts)
 		return nil
 	},
 	Version: version.Version,
